requirements.txt
# Core RAG & LLM
llama-index>=0.10.0,<0.11.0
llama-index-vector-stores-qdrant
llama-index-embeddings-ollama
llama-index-llms-ollama
qdrant-client>=1.9.0,<2.0.0
ollama>=0.2.0,<1.0.0

# Ingestion
feedparser>=6.0.0,<7.0.0
trafilatura>=1.9.0,<2.0.0
redis>=5.0.0,<6.0.0
celery>=5.3.0,<6.0.0
httpx>=0.27.0,<1.0.0
playwright>=1.40.0

# NLP Processing
spacy>=3.7.0,<4.0.0
transformers>=4.40.0,<5.0.0
torch>=2.3.0
sentencepiece>=0.2.0

# Market Data & Historical
yfinance>=0.2.0,<1.0.0
requests>=2.31.0,<3.0.0
openai>=1.0.0

# API
fastapi>=0.111.0,<1.0.0
uvicorn>=0.30.0,<1.0.0
pydantic>=2.7.0,<3.0.0
pydantic-settings>=2.0.0

# Monitoring
prometheus-client>=0.20.0
prometheus-fastapi-instrumentator>=7.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
structlog>=24.0.0
